% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CEOdata-package.R
\docType{package}
\encoding{UTF-8}
\name{CEOdata-package}
\alias{CEOdata-package}
\title{CEOdata: CEO microdata access from the Generalitat open data platform}
\description{
Provides easy and convenient access to microdata from the Centre d'Estudis d'Opinio (CEO).
Datasets are retrieved from the open data platform of the Generalitat de Catalunya
and returned as tidy tibbles.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://ceo.gencat.cat/ca/inici/}
  \item \url{https://github.com/ceopinio/CEOdata/}
  \item Report bugs at \url{https://github.com/ceopinio/CEOdata/issues/}
}

}
\author{
\strong{Maintainer}: Joel Ardiaca \email{joelardiacaarnau@gencat.cat} (\href{https://orcid.org/0009-0004-4937-8941}{ORCID})

Authors:
\itemize{
  \item Xavier Fernández-i-Marín \email{xavier.fim@gmail.com} (\href{https://orcid.org/0000-0002-9522-8870}{ORCID})
}

}
\keyword{internal}
