% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{ceodata_download_and_read}
\alias{ceodata_download_and_read}
\title{Download and read a dataset from a URL (.sav or .zip containing a single .sav)}
\usage{
ceodata_download_and_read(url, raw = FALSE)
}
\arguments{
\item{url}{Character scalar. URL to .sav or .zip.}

\item{raw}{Logical. If TRUE, do not convert labelled vectors.}
}
\value{
A data.frame/tibble.
}
\description{
Download and read a dataset from a URL (.sav or .zip containing a single .sav)
}
\keyword{internal}
