% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CEOaccumulated_meta.R
\name{CEOaccumulated_metadata}
\alias{CEOaccumulated_metadata}
\title{Internal cache for accumulated-series metadata}
\usage{
CEOaccumulated_metadata()
}
\description{
Internal cache for accumulated-series metadata
}
\keyword{internal}
