% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{ceodata_detect_type}
\alias{ceodata_detect_type}
\title{Detect input type from a path or URL}
\usage{
ceodata_detect_type(x)
}
\arguments{
\item{x}{Character scalar. File path or URL.}
}
\value{
One of "sav" or "zip".
}
\description{
Detect input type from a path or URL
}
\keyword{internal}
