% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CEOaccumulated_meta.R
\name{getCEOaccumulated_metadata}
\alias{getCEOaccumulated_metadata}
\title{Download and prepare the metadata of accumulated microdata series (Dades Obertes)}
\usage{
getCEOaccumulated_metadata()
}
\value{
A tibble with the accumulated series metadata.
}
\description{
Source dataset (Socrata):
https://analisi.transparenciacatalunya.cat/resource/gp4k-sxxn.json
}
\keyword{internal}
