% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CEOmeta.R
\encoding{UTF-8}
\name{read.ceo.json}
\alias{read.ceo.json}
\title{Internal function to properly read the JSON from CEO}
\usage{
read.ceo.json(x)
}
\arguments{
\item{x}{JSON data structure}
}
\description{
Used to address limitations of the JSON format provided.
}
\keyword{internal}
