% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CEOmeta.R
\docType{data}
\encoding{UTF-8}
\name{the}
\alias{the}
\title{Internal function to get the metadata of CEO surveys into cache}
\format{
An object of class \code{environment} of length 0.
}
\usage{
the
}
\description{
Used when loading the package. It gets the latest metadata of CEO surveys,
cleans it and makes it ready for the rest of functions in the package.
}
\keyword{internal}
